Applied aggregation/encapsulation for the sources. And I made url/data_source class attributes instead of static methods
