+++ COVID19Py/__init__.py
@@ -1,1 +1,2 @@
 from .covid19 import COVID19
+from .source import Source
