+++ COVID19Py/covid19.py
@@ -1,53 +1,18 @@
 from typing import Dict, List
 import requests
 import json
-
+from source import Source
 class COVID19(object):
     default_url = "https://covid-tracker-us.herokuapp.com"
-    url = ""
-    data_source = ""
     previousData = None
     latestData = None
-    _valid_data_sources = []
-
-    mirrors_source = "https://raw.github.com/Kamaropoulos/COVID19Py/master/mirrors.json"
-    mirrors = None
-
     def __init__(self, url="https://covid-tracker-us.herokuapp.com", data_source='jhu'):
+        self.data_source = data_source
+        self.url = url
         # Skip mirror checking if custom url was passed
         if url == self.default_url:
-            # Load mirrors
-            response = requests.get(self.mirrors_source)
-            response.raise_for_status()
-            self.mirrors = response.json()
-
-            # Try to get sources as a test
-            for mirror in self.mirrors:
-                # Set URL of mirror
-                self.url = mirror["url"]
-                result = None
-                try:
-                    result = self._getSources()
-                except Exception as e:
-                    # URL did not work, reset it and move on
-                    self.url = ""
-                    continue
-
-                # TODO: Should have a better health-check, this is way too hacky...
-                if "jhu" in result:
-                    # We found a mirror that worked just fine, let's stick with it
-                    break
-
-                # None of the mirrors worked. Raise an error to inform the user.
-                raise RuntimeError("No available API mirror was found.")
-
-        else:
-            self.url = url
-
-        self._valid_data_sources = self._getSources()
-        if data_source not in self._valid_data_sources:
-            raise ValueError("Invalid data source. Expected one of: %s" % self._valid_data_sources)
-        self.data_source = data_source
+            self.url = Source(self.url,data_source).getUrl()
+            
 
     def _update(self, timelines):
         latest = self.getLatest()
@@ -59,10 +24,7 @@ class COVID19(object):
             "locations": locations
         }
 
-    def _getSources(self):
-        response = requests.get(self.url + "/v2/sources")
-        response.raise_for_status()
-        return response.json()["sources"]
+    
 
     def _request(self, endpoint, params=None):
         if params is None:
